refactor: move models do domain directory
