+++ src/app/core/services/vehicle-details.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@angular/core';
 import {environment} from '@environments/environment.development';
 import {HttpClient} from '@angular/common/http';
 import {Observable} from 'rxjs';
-import {VehicleDetails} from '@models/vehicledetails';
+import {VehicleDetails} from '@domain/vehicle/vehicledetails';
 
 @Injectable({
   providedIn: 'root'
